refactor(server,types): carve file-size headroom for the handler batches
Batch 1, PR 5 of 5. No behavior change — pure moves.

  server/event_enrollment/tests.rs  696 -> 7 files by algorithm, max 214
  server/requests.rs                695 -> mod 547 + unconfirmed 152
  server/event_notifications.rs     617 -> 281 + tests 333
  handlers/alarm_event.rs           240 -> 6 files, one per service, max 106
  types event_parameter/mod.rs      630 -> mod 260 + codec 216 + fault 165

event_enrollment/tests.rs had 4 lines of headroom and Batch 2 is entirely
Event Enrollment work, so it went first. alarm_event.rs was nowhere near the
cap at 240, but Batch 4 lands seven handler PRs in it; one file per service
means those seven stop serializing on a single module and stop racing the cap.

Visibility is preserved, not widened:

- handle_unconfirmed_request was `pub(super)` in server/requests.rs, meaning
  visible in `server`. From one level deeper it is `pub(in crate::server)` —
  the same scope stated explicitly, so server::dispatch still reaches it.
- The event-parameter codec helpers were private to event_parameter/mod.rs and
  are now `pub(super)` in codec.rs, which is the same module scope.

Two mistakes worth recording, both caught by the compiler:

- Slicing alarm_event.rs at each `pub fn` line orphaned every `///` block at
  the tail of the previous file. Boundaries now walk back over the doc comment.
- The trailing `#[cfg(test)] mod tests;` was swept into the event_parameter
  fault segment, where it resolved to a nonexistent fault/tests.rs.

Also removed the unused imports the duplicated header blocks introduced, in
this split and one left over from #197.

Verified: 2234 tests exit 0; clippy exit 0 with no unused-import warnings; fmt
clean; cap gate green; leaf test-function names identical to dev, so no test
was added, dropped, or renamed.

crates/bacnet-cli/src/main.rs is also at exactly 700/700 but is out of scope —
no batch touches the CLI, and extracting its clap surface needs visibility and
import work rather than a pure move. Filed as #198.
