+++ crates/bacnet-server/src/handlers/alarm_event/mod.rs
@@ -1,0 +1,15 @@
+//! Alarm and event service handlers (ASHRAE 135-2020 Clause 13).
+
+mod acknowledge_alarm;
+mod get_alarm_summary;
+mod get_enrollment_summary;
+mod get_event_information;
+mod life_safety_operation;
+mod text_message;
+
+pub use acknowledge_alarm::*;
+pub use get_alarm_summary::*;
+pub use get_enrollment_summary::*;
+pub use get_event_information::*;
+pub use life_safety_operation::*;
+pub use text_message::*;
