+++ crates/bacnet-server/src/server/mod.rs
@@ -776,6 +776,8 @@ mod segmentation;
 #[cfg(test)]
 mod cov_notifications_tests;
 #[cfg(test)]
+mod event_notifications_tests;
+#[cfg(test)]
 mod segmentation_tests;
 #[cfg(test)]
 mod tests;
