+++ crates/bacnet-types/src/constructed/event_parameter/mod.rs
@@ -350,426 +350,9 @@ impl BACnetEventParameter {
     }
 }
 
-// ---- encode/decode helpers ----
-
-/// Build a [`PropertyValue::BitString`] from an `(unused_bits, data)` pair.
-fn bitstring_pv((unused_bits, data): (u8, Vec<u8>)) -> PropertyValue {
-    PropertyValue::BitString { unused_bits, data }
-}
-
-/// Build a [`PropertyValue`] for a [`BACnetPropertyStates`] tag.
-///
-/// `BACnetPropertyStates` is itself a CHOICE; absent full context-tagged
-/// framing we carry its discriminant as the low byte and the raw payload as an
-/// octet string. This round-trips through the flat-`List` encoding.
-pub(super) fn property_state_pv(state: &BACnetPropertyStates) -> PropertyValue {
-    let (tag, data) = property_state_parts(state);
-    PropertyValue::List(vec![
-        PropertyValue::Unsigned(tag as u64),
-        PropertyValue::OctetString(data),
-    ])
-}
-
-/// Extract `(tag, raw data)` from a [`BACnetPropertyStates`].
-fn property_state_parts(state: &BACnetPropertyStates) -> (u8, Vec<u8>) {
-    match state {
-        BACnetPropertyStates::BooleanValue(v) => (0u8, vec![u8::from(*v)]),
-        BACnetPropertyStates::BinaryValue(v) => (1, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::EventType(v) => (2, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::Polarity(v) => (3, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::ProgramChange(v) => (4, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::ProgramState(v) => (5, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::ReasonForHalt(v) => (6, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::Reliability(v) => (7, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::State(v) => (8, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::SystemStatus(v) => (9, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::Units(v) => (10, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::UnsignedValue(v) => (11, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::LifeSafetyMode(v) => (12, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::LifeSafetyState(v) => (13, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::DoorAlarmState(v) => (14, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::Action(v) => (15, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::DoorSecuredStatus(v) => (16, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::DoorStatus(v) => (17, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::DoorValue(v) => (18, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::LiftCarDirection(v) => (40, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::LiftCarDoorCommand(v) => (42, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::TimerState(v) => (38, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::TimerTransition(v) => (39, v.to_le_bytes().to_vec()),
-        BACnetPropertyStates::Other { tag, data } => (*tag, data.clone()),
-    }
-}
-
-/// Build a [`PropertyValue`] for a [`BACnetDeviceObjectPropertyReference`].
-pub(super) fn device_object_property_reference_pv(
-    r: &BACnetDeviceObjectPropertyReference,
-) -> PropertyValue {
-    PropertyValue::List(vec![
-        PropertyValue::ObjectIdentifier(r.object_identifier),
-        PropertyValue::Unsigned(r.property_identifier as u64),
-        match r.property_array_index {
-            Some(idx) => PropertyValue::Unsigned(idx as u64),
-            None => PropertyValue::Null,
-        },
-        match r.device_identifier {
-            Some(dev) => PropertyValue::ObjectIdentifier(dev),
-            None => PropertyValue::Null,
-        },
-    ])
-}
-
-/// Read the next [`PropertyValue`] at `idx`, advancing it.
-fn take_pv<'a>(items: &'a [PropertyValue], idx: &mut usize) -> Option<&'a PropertyValue> {
-    let v = items.get(*idx)?;
-    *idx += 1;
-    Some(v)
-}
-
-/// Read a `Unsigned` as a `u32`.
-fn take_u32(items: &[PropertyValue], idx: &mut usize) -> Result<u32, Error> {
-    match items.get(*idx) {
-        Some(PropertyValue::Unsigned(v)) => {
-            *idx += 1;
-            u32::try_from(*v).map_err(|_| Error::decoding(*idx, "Unsigned exceeds u32"))
-        }
-        _ => Err(Error::decoding(*idx, "expected Unsigned")),
-    }
-}
-
-/// Read a `Unsigned` as a `u16`.
-fn take_u16(items: &[PropertyValue], idx: &mut usize) -> Result<u16, Error> {
-    match items.get(*idx) {
-        Some(PropertyValue::Unsigned(v)) => {
-            *idx += 1;
-            u16::try_from(*v).map_err(|_| Error::decoding(*idx, "Unsigned exceeds u16"))
-        }
-        _ => Err(Error::decoding(*idx, "expected Unsigned")),
-    }
-}
-
-/// Read a `Real`/`Double`/`Unsigned` as `f32`.
-fn take_real(items: &[PropertyValue], idx: &mut usize) -> Result<f32, Error> {
-    let v = match items.get(*idx) {
-        Some(PropertyValue::Real(v)) => *v,
-        Some(PropertyValue::Double(v)) => *v as f32,
-        Some(PropertyValue::Unsigned(v)) => *v as f32,
-        _ => return Err(Error::decoding(*idx, "expected Real")),
-    };
-    *idx += 1;
-    Ok(v)
-}
-
-/// Read a bitstring `(unused_bits, data)` pair.
-fn take_bitstring(items: &[PropertyValue], idx: &mut usize) -> Result<(u8, Vec<u8>), Error> {
-    match items.get(*idx) {
-        Some(PropertyValue::BitString { unused_bits, data }) => {
-            *idx += 1;
-            Ok((*unused_bits, data.clone()))
-        }
-        _ => Err(Error::decoding(*idx, "expected BitString")),
-    }
-}
-
-/// Read a `SEQUENCE OF BIT STRING` as a list of `(unused_bits, data)` pairs.
-fn take_bitstring_list(
-    items: &[PropertyValue],
-    idx: &mut usize,
-) -> Result<Vec<(u8, Vec<u8>)>, Error> {
-    match items.get(*idx) {
-        Some(PropertyValue::List(inner)) => {
-            *idx += 1;
-            inner
-                .iter()
-                .map(|v| take_bitstring(core::slice::from_ref(v), &mut 0))
-                .collect()
-        }
-        _ => Err(Error::decoding(*idx, "expected list of bitstrings")),
-    }
-}
-
-/// Read a [`BACnetDeviceObjectPropertyReference`].
-pub(super) fn take_dopr(
-    items: &[PropertyValue],
-    idx: &mut usize,
-) -> Result<BACnetDeviceObjectPropertyReference, Error> {
-    let Some(PropertyValue::List(inner)) = items.get(*idx) else {
-        return Err(Error::decoding(
-            *idx,
-            "expected device object property reference",
-        ));
-    };
-    *idx += 1;
-    if inner.len() < 4 {
-        return Err(Error::decoding(
-            *idx,
-            "device object property reference too short",
-        ));
-    }
-    let object_identifier = match &inner[0] {
-        PropertyValue::ObjectIdentifier(oid) => *oid,
-        _ => return Err(Error::decoding(0, "reference object id missing")),
-    };
-    let property_identifier = match &inner[1] {
-        PropertyValue::Unsigned(v) => *v as u32,
-        _ => return Err(Error::decoding(1, "reference property id missing")),
-    };
-    let property_array_index = match &inner[2] {
-        PropertyValue::Unsigned(v) => Some(*v as u32),
-        PropertyValue::Null => None,
-        _ => return Err(Error::decoding(2, "reference array index invalid")),
-    };
-    let device_identifier = match &inner[3] {
-        PropertyValue::ObjectIdentifier(dev) => Some(*dev),
-        PropertyValue::Null => None,
-        _ => return Err(Error::decoding(3, "reference device id invalid")),
-    };
-    Ok(BACnetDeviceObjectPropertyReference {
-        object_identifier,
-        property_identifier,
-        property_array_index,
-        device_identifier,
-    })
-}
-
-/// Read a `SEQUENCE OF BACnetPropertyStates`.
-pub(super) fn take_property_state_list(
-    items: &[PropertyValue],
-    idx: &mut usize,
-) -> Result<Vec<BACnetPropertyStates>, Error> {
-    let Some(PropertyValue::List(inner)) = items.get(*idx) else {
-        return Err(Error::decoding(*idx, "expected list of property states"));
-    };
-    *idx += 1;
-    inner.iter().map(property_state_from_pv).collect()
-}
-
-/// Reconstruct a [`BACnetPropertyStates`] from its flat-`List` encoding.
-pub(super) fn property_state_from_pv(pv: &PropertyValue) -> Result<BACnetPropertyStates, Error> {
-    let PropertyValue::List(inner) = pv else {
-        return Err(Error::decoding(0, "property state is not a List"));
-    };
-    let Some((tag_pv, rest)) = inner.split_first() else {
-        return Err(Error::decoding(0, "property state list empty"));
-    };
-    let PropertyValue::Unsigned(tag) = tag_pv else {
-        return Err(Error::decoding(0, "property state tag not Unsigned"));
-    };
-    let data = match rest.first() {
-        Some(PropertyValue::OctetString(b)) => b.clone(),
-        _ => return Err(Error::decoding(1, "property state data not octets")),
-    };
-    let read_u32 = |data: &[u8]| -> Result<u32, Error> {
-        data.try_into()
-            .map(u32::from_le_bytes)
-            .map_err(|_| Error::decoding(1, "property state data wrong length"))
-    };
-    Ok(match *tag as u8 {
-        0 => BACnetPropertyStates::BooleanValue(data.first().copied().unwrap_or(0) != 0),
-        1 => BACnetPropertyStates::BinaryValue(read_u32(&data)?),
-        2 => BACnetPropertyStates::EventType(read_u32(&data)?),
-        3 => BACnetPropertyStates::Polarity(read_u32(&data)?),
-        4 => BACnetPropertyStates::ProgramChange(read_u32(&data)?),
-        5 => BACnetPropertyStates::ProgramState(read_u32(&data)?),
-        6 => BACnetPropertyStates::ReasonForHalt(read_u32(&data)?),
-        7 => BACnetPropertyStates::Reliability(read_u32(&data)?),
-        8 => BACnetPropertyStates::State(read_u32(&data)?),
-        9 => BACnetPropertyStates::SystemStatus(read_u32(&data)?),
-        10 => BACnetPropertyStates::Units(read_u32(&data)?),
-        11 => BACnetPropertyStates::UnsignedValue(read_u32(&data)?),
-        12 => BACnetPropertyStates::LifeSafetyMode(read_u32(&data)?),
-        13 => BACnetPropertyStates::LifeSafetyState(read_u32(&data)?),
-        14 => BACnetPropertyStates::DoorAlarmState(read_u32(&data)?),
-        15 => BACnetPropertyStates::Action(read_u32(&data)?),
-        16 => BACnetPropertyStates::DoorSecuredStatus(read_u32(&data)?),
-        17 => BACnetPropertyStates::DoorStatus(read_u32(&data)?),
-        18 => BACnetPropertyStates::DoorValue(read_u32(&data)?),
-        40 => BACnetPropertyStates::LiftCarDirection(read_u32(&data)?),
-        42 => BACnetPropertyStates::LiftCarDoorCommand(read_u32(&data)?),
-        38 => BACnetPropertyStates::TimerState(read_u32(&data)?),
-        39 => BACnetPropertyStates::TimerTransition(read_u32(&data)?),
-        other => BACnetPropertyStates::Other { tag: other, data },
-    })
-}
-
-// ---------------------------------------------------------------------------
-// FaultParameters structured round trip (Clause 12.12.50 -- Fault_Parameters)
-// ---------------------------------------------------------------------------
-
-/// Variant tag carried as the leading element of the flat-`List` encoding.
-mod fault_parameter_tag {
-    pub const NONE: u8 = 0;
-    pub const CHARACTER_STRING: u8 = 1;
-    pub const EXTENDED: u8 = 2;
-    pub const LIFE_SAFETY: u8 = 3;
-    pub const STATE: u8 = 4;
-    pub const STATUS_FLAGS: u8 = 5;
-    pub const OUT_OF_RANGE: u8 = 6;
-    pub const LISTED: u8 = 7;
-}
-
-impl crate::constructed::FaultParameters {
-    /// Encode this fault-parameter set as a flat [`PropertyValue::List`].
-    ///
-    /// The first element is the variant tag as [`PropertyValue::Unsigned`],
-    /// followed by the variant's fields in declaration order. This is the
-    /// on-the-wire representation used by an EventEnrollment object's
-    /// `Fault_Parameters` property; see [`Self::decode_property_value`].
-    pub fn encode_property_value(&self) -> PropertyValue {
-        use crate::constructed::FaultParameters as F;
-        match self {
-            F::FaultNone => PropertyValue::List(vec![PropertyValue::Unsigned(
-                fault_parameter_tag::NONE as u64,
-            )]),
-            F::FaultCharacterString { fault_values } => PropertyValue::List(vec![
-                PropertyValue::Unsigned(fault_parameter_tag::CHARACTER_STRING as u64),
-                PropertyValue::List(
-                    fault_values
-                        .iter()
-                        .map(|s| PropertyValue::CharacterString(s.clone()))
-                        .collect(),
-                ),
-            ]),
-            F::FaultExtended {
-                vendor_id,
-                extended_fault_type,
-                parameters,
-            } => PropertyValue::List(vec![
-                PropertyValue::Unsigned(fault_parameter_tag::EXTENDED as u64),
-                PropertyValue::Unsigned(*vendor_id as u64),
-                PropertyValue::Unsigned(*extended_fault_type as u64),
-                PropertyValue::OctetString(parameters.clone()),
-            ]),
-            F::FaultLifeSafety {
-                fault_values,
-                mode_for_reference,
-            } => PropertyValue::List(vec![
-                PropertyValue::Unsigned(fault_parameter_tag::LIFE_SAFETY as u64),
-                PropertyValue::List(
-                    fault_values
-                        .iter()
-                        .map(|v| PropertyValue::Unsigned(*v as u64))
-                        .collect(),
-                ),
-                device_object_property_reference_pv(mode_for_reference),
-            ]),
-            F::FaultState { fault_values } => PropertyValue::List(vec![
-                PropertyValue::Unsigned(fault_parameter_tag::STATE as u64),
-                PropertyValue::List(fault_values.iter().map(property_state_pv).collect()),
-            ]),
-            F::FaultStatusFlags { reference } => PropertyValue::List(vec![
-                PropertyValue::Unsigned(fault_parameter_tag::STATUS_FLAGS as u64),
-                device_object_property_reference_pv(reference),
-            ]),
-            F::FaultOutOfRange {
-                min_normal,
-                max_normal,
-            } => PropertyValue::List(vec![
-                PropertyValue::Unsigned(fault_parameter_tag::OUT_OF_RANGE as u64),
-                PropertyValue::Double(*min_normal),
-                PropertyValue::Double(*max_normal),
-            ]),
-            F::FaultListed { reference } => PropertyValue::List(vec![
-                PropertyValue::Unsigned(fault_parameter_tag::LISTED as u64),
-                device_object_property_reference_pv(reference),
-            ]),
-        }
-    }
-
-    /// Decode a [`PropertyValue`] previously produced by
-    /// [`Self::encode_property_value`].
-    pub fn decode_property_value(value: &PropertyValue) -> Result<Self, Error> {
-        use crate::constructed::FaultParameters as F;
-        let PropertyValue::List(items) = value else {
-            return Err(Error::decoding(0, "Fault_Parameters is not a List"));
-        };
-        let Some((tag_pv, rest)) = items.split_first() else {
-            return Err(Error::decoding(0, "Fault_Parameters list is empty"));
-        };
-        let PropertyValue::Unsigned(tag) = tag_pv else {
-            return Err(Error::decoding(0, "Fault_Parameters tag is not Unsigned"));
-        };
-        let mut idx = 0;
-        Ok(match *tag as u8 {
-            fault_parameter_tag::NONE => F::FaultNone,
-            fault_parameter_tag::CHARACTER_STRING => {
-                let Some(PropertyValue::List(inner)) = rest.get(idx) else {
-                    return Err(Error::decoding(idx, "fault characterstring values missing"));
-                };
-                idx += 1;
-                let fault_values = inner
-                    .iter()
-                    .map(|v| match v {
-                        PropertyValue::CharacterString(s) => Ok(s.clone()),
-                        _ => Err(Error::decoding(idx, "fault value not a string")),
-                    })
-                    .collect::<Result<Vec<String>, Error>>()?;
-                F::FaultCharacterString { fault_values }
-            }
-            fault_parameter_tag::EXTENDED => {
-                let vendor_id = take_u16(rest, &mut idx)?;
-                let extended_fault_type = take_u32(rest, &mut idx)?;
-                let parameters = match take_pv(rest, &mut idx) {
-                    Some(PropertyValue::OctetString(b)) => b.clone(),
-                    _ => return Err(Error::decoding(idx, "extended fault params not octets")),
-                };
-                F::FaultExtended {
-                    vendor_id,
-                    extended_fault_type,
-                    parameters,
-                }
-            }
-            fault_parameter_tag::LIFE_SAFETY => {
-                let Some(PropertyValue::List(inner)) = rest.get(idx) else {
-                    return Err(Error::decoding(idx, "life-safety fault values missing"));
-                };
-                idx += 1;
-                let fault_values = inner
-                    .iter()
-                    .map(|v| match v {
-                        PropertyValue::Unsigned(u) => Ok(*u as u32),
-                        _ => Err(Error::decoding(idx, "life-safety value not unsigned")),
-                    })
-                    .collect::<Result<Vec<u32>, Error>>()?;
-                let mode_for_reference = take_dopr(rest, &mut idx)?;
-                F::FaultLifeSafety {
-                    fault_values,
-                    mode_for_reference,
-                }
-            }
-            fault_parameter_tag::STATE => {
-                let list_of_values = take_property_state_list(rest, &mut idx)?;
-                F::FaultState {
-                    fault_values: list_of_values,
-                }
-            }
-            fault_parameter_tag::STATUS_FLAGS => {
-                let reference = take_dopr(rest, &mut idx)?;
-                F::FaultStatusFlags { reference }
-            }
-            fault_parameter_tag::OUT_OF_RANGE => {
-                let min_normal = match take_pv(rest, &mut idx) {
-                    Some(PropertyValue::Double(v)) => *v,
-                    Some(PropertyValue::Real(v)) => *v as f64,
-                    _ => return Err(Error::decoding(idx, "min_normal not a double")),
-                };
-                let max_normal = match take_pv(rest, &mut idx) {
-                    Some(PropertyValue::Double(v)) => *v,
-                    Some(PropertyValue::Real(v)) => *v as f64,
-                    _ => return Err(Error::decoding(idx, "max_normal not a double")),
-                };
-                F::FaultOutOfRange {
-                    min_normal,
-                    max_normal,
-                }
-            }
-            fault_parameter_tag::LISTED => {
-                let reference = take_dopr(rest, &mut idx)?;
-                F::FaultListed { reference }
-            }
-            other => return Err(Error::decoding(0, format!("unknown fault tag {other}"))),
-        })
-    }
-}
+mod codec;
+mod fault;
+use codec::*;
 
 #[cfg(test)]
 mod tests;
